+++ sdk/metric/reader_test.go
@@ -157,7 +157,7 @@ func (ts *readerTestSuite) TestSDKFailureBlocksExternalProducer() {
 	ts.Equal(metricdata.ResourceMetrics{}, m)
 }
 
-func (ts *readerTestSuite) TestMethodConcurrency() {
+func (ts *readerTestSuite) TestMethodConcurrentSafe() {
 	// Requires the race-detector (a default test option for the project).
 
 	// All reader methods should be concurrent-safe.
@@ -168,6 +168,18 @@ func (ts *readerTestSuite) TestMethodConcurrency() {
 	var wg sync.WaitGroup
 	const threads = 2
 	for i := 0; i < threads; i++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			_ = ts.Reader.temporality(InstrumentKindCounter)
+		}()
+
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			_ = ts.Reader.aggregation(InstrumentKindCounter)
+		}()
+
 		wg.Add(1)
 		go func() {
 			defer wg.Done()